# backend/docs/swagger.yaml
basePath: /
definitions:
  handlers.eventRequest:
    properties:
      capacity:
        type: integer
      description:
        type: string
      duration:
        type: number
      location:
        type: string
      name:
        type: string
      startTimestamp:
        format: date-time
        type: string
    type: object
  handlers.healthResponse:
    properties:
      checks:
        additionalProperties:
          type: string
        type: object
      status:
        type: string
    type: object
  handlers.participantRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        example: min 8 characters
        type: string
    type: object
  models.Event:
    properties:
      capacity:
        type: integer
      description:
        type: string
      duration:
        type: number
      id:
        type: string
      location:
        type: string
      name:
        type: string
      participants:
        items:
          $ref: '#/definitions/models.Participant'
        type: array
      startTimestamp:
        type: string
    type: object
  models.Participant:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  services.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  services.TokenResponse:
    properties:
      token:
        type: string
    type: object
info:
  contact: {}
  description: Events management API
  title: VB-Events API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Returns a signed JWT valid for 24 hours
      parameters:
      - description: Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.TokenResponse'
        "400":
          description: Invalid request body
          schema:
            type: object
        "401":
          description: Invalid credentials
          schema:
            type: object
        "500":
          description: Internal error
          schema:
            type: object
      summary: Log in with email and password
      tags:
      - auth
  /auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Participant'
        "401":
          description: Not authenticated
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get the currently authenticated participant
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a participant account with a name, unique email and password
      parameters:
      - description: Participant to register
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.participantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Participant'
        "400":
          description: Invalid body, or missing name/email/password (min 8 chars)
          schema:
            type: object
        "409":
          description: Email already in use
          schema:
            type: object
        "500":
          description: Internal error
          schema:
            type: object
      summary: Register a new participant
      tags:
      - auth
  /events:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Event'
            type: array
        "500":
          description: Internal error
          schema:
            type: object
      summary: List all events
      tags:
      - events
    post:
      consumes:
      - application/json
      description: Creates an event with a name and start timestamp
      parameters:
      - description: Event to create
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.eventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Event'
        "400":
          description: Invalid body, or missing name/startTimestamp
          schema:
            type: object
        "500":
          description: Internal error
          schema:
            type: object
      summary: Create a new event
      tags:
      - events
  /events/{id}:
    delete:
      parameters:
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No content
        "400":
          description: Invalid event id
          schema:
            type: object
        "404":
          description: Event not found
          schema:
            type: object
        "500":
          description: Internal error
          schema:
            type: object
      summary: Delete an event
      tags:
      - events
    get:
      parameters:
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Event'
        "400":
          description: Invalid event id
          schema:
            type: object
        "404":
          description: Event not found
          schema:
            type: object
        "500":
          description: Internal error
          schema:
            type: object
      summary: Get an event by ID with its participants
      tags:
      - events
    put:
      consumes:
      - application/json
      parameters:
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated event data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.eventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Event'
        "400":
          description: Invalid body or id, or missing name/startTimestamp
          schema:
            type: object
        "404":
          description: Event not found
          schema:
            type: object
        "500":
          description: Internal error
          schema:
            type: object
      summary: Update an event
      tags:
      - events
  /events/{id}/participants:
    post:
      description: Adds the authenticated participant to the event if there is capacity
        remaining and they are not already registered
      parameters:
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No content
        "400":
          description: Invalid event id
          schema:
            type: object
        "401":
          description: Not authenticated
          schema:
            type: object
        "404":
          description: Event not found
          schema:
            type: object
        "409":
          description: Already registered, or event is at capacity
          schema:
            type: object
        "500":
          description: Internal error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Register the current user as a participant of an event
      tags:
      - events
  /events/after:
    get:
      parameters:
      - description: RFC3339 timestamp
        in: query
        name: timestamp
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Event'
            type: array
        "400":
          description: Invalid or missing timestamp
          schema:
            type: object
        "500":
          description: Internal error
          schema:
            type: object
      summary: Find events starting after a timestamp
      tags:
      - events
  /events/before:
    get:
      parameters:
      - description: RFC3339 timestamp
        in: query
        name: timestamp
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Event'
            type: array
        "400":
          description: Invalid or missing timestamp
          schema:
            type: object
        "500":
          description: Internal error
          schema:
            type: object
      summary: Find events starting before a timestamp
      tags:
      - events
  /events/between:
    get:
      parameters:
      - description: RFC3339 start timestamp
        in: query
        name: start
        required: true
        type: string
      - description: RFC3339 end timestamp
        in: query
        name: end
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Event'
            type: array
        "400":
          description: Invalid or missing timestamps
          schema:
            type: object
        "500":
          description: Internal error
          schema:
            type: object
      summary: Find events starting between two timestamps
      tags:
      - events
  /events/name/{name}:
    get:
      parameters:
      - description: Event name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Event'
        "404":
          description: Event not found
          schema:
            type: object
        "500":
          description: Internal error
          schema:
            type: object
      summary: Find an event by exact name
      tags:
      - events
  /events/search:
    get:
      parameters:
      - description: Search term
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Event'
            type: array
        "400":
          description: Missing search term
          schema:
            type: object
        "500":
          description: Internal error
          schema:
            type: object
      summary: Search events by partial name, description, or location
      tags:
      - events
  /healthz:
    get:
      description: Returns service status and verifies database connectivity
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.healthResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/handlers.healthResponse'
      summary: Service health check
      tags:
      - health
  /participants:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Participant'
            type: array
        "500":
          description: Internal error
          schema:
            type: object
      summary: List all participants
      tags:
      - participants
  /participants/{id}:
    delete:
      parameters:
      - description: Participant ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No content
        "400":
          description: Invalid participant id
          schema:
            type: object
        "404":
          description: Participant not found
          schema:
            type: object
        "500":
          description: Internal error
          schema:
            type: object
      summary: Delete a participant
      tags:
      - participants
    get:
      parameters:
      - description: Participant ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Participant'
        "400":
          description: Invalid participant id
          schema:
            type: object
        "404":
          description: Participant not found
          schema:
            type: object
        "500":
          description: Internal error
          schema:
            type: object
      summary: Get a participant by ID
      tags:
      - participants
    put:
      consumes:
      - application/json
      parameters:
      - description: Participant ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated participant data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.participantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Participant'
        "400":
          description: Invalid body or id, or missing name/email
          schema:
            type: object
        "404":
          description: Participant not found
          schema:
            type: object
        "409":
          description: Email already in use
          schema:
            type: object
        "500":
          description: Internal error
          schema:
            type: object
      summary: Update a participant
      tags:
      - participants
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"